feat: implement education screen UI
Implemented the UI for the Education screen using Bloc for state management.
Added validation for input fields and a button to proceed to the next step.
Created EducationEntity to represent education details and updated app routing to start with the Education screen.
Refactored the widget structure to improve organization and maintainability, and moved widgets into a new 'user info' directory.
Updated the Profile screen to reflect the changes.