feat: implement project list widget and refactor project screen
Implemented a `ProjectList` widget to display a list of projects with details and delete/undo functionality.
Refactored the `ProjectScreen` to use the `ProjectList` widget instead of inline list building.
Updated `ProjectViewmodel` to remove unnecessary listeners.
Modified `ShortPath` to start with the Project screen.
Moved the project list display logic from the `ProjectScreen` to the `ProjectList` widget for better organization and reusability.
Added the ability to remove a project from the list with a confirmation snackbar and an undo option.